requirements.txt
asgiref==3.6.0
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4

Django==4.1.6
django-filter==22.1
django-rosetta==0.9.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5

Flask==2.2.3
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
Markdown==3.4.1
MarkupSafe==2.1.2
packaging==23.0
Pillow==9.4.0
polib==1.1.1
pycodestyle==2.10.0
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
swagger-spec-validator==3.0.3
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
Werkzeug==2.2.3
